refactor(runs): render comparison HTML client-side from static template
Review feedback on #46: string-building the page in TS template literals
left the embedded JSON unescaped against </script> breakout and made the
word-diff algorithm unlintable and untestable.

html.ts now just inlines client.js into template.html and embeds the
ComparisonData as a JSON blob with < escaped to \u003c — the single
server-side escaping site. All rendering happens client-side in
client.js, a real file with unit tests for wordDiff and escaping.

The build copies template.html to dist; client.js is emitted by tsc
(allowJs). Shared test fixtures move to fixtures.ts.
