+++ package.json
@@ -4,16 +4,16 @@
   "exports": "./dist/server.js",
   "type": "module",
   "scripts": {
-    "build": "tsc",
+    "build": "tsc && cp src/lib/runs/template.html dist/lib/runs/",
     "dev": "./scripts/dev.sh",
     "start": "node dist/server.js",
     "cli": "node dist/cli.js",
-    "cli:dev": "tsc && node dist/cli.js",
+    "cli:dev": "npm run build && node dist/cli.js",
     "test": "vitest run",
     "typecheck": "tsc --noEmit",
     "test:watch": "vitest",
     "test:upload": "curl -X POST http://localhost:3005/dev/test-upload",
-    "smoke": "tsc && node dist/cli.js smoke"
+    "smoke": "npm run build && node dist/cli.js smoke"
   },
   "dependencies": {
     "@anthropic-ai/sdk": "^0.71.2",
